+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Removed `GenericID`, most ID's are of type `string` now
 - Made typing of some objects more strict
 - Various other bugfixes and improvments
+- Removed legacy `iptv` action
 
 ## v2.0.0